refactor: Update Prisma schema to use UUIDs for ID fields
This commit updates the Prisma schema to use UUIDs for the ID fields of the `User` and `Course` models. The `id` fields are now defined as `String @id @default(uuid())`. This change improves the uniqueness and security of the ID values in the database.

Recent user commits:
- fix: updated sign up screen

Recent repository commits:
- refactor: Update Prisma database provider to MongoDB
- refactor: Update database provider to MongoDB in Prisma schema
- feat: Add @prisma/client and mongodb dependencies