refactor: Update sign-up screen to use "/auth/sign-up" as the href
The code change in the file "src/app-components/hero.tsx" updates the href value of a Link component in the HeroSection. The href is changed from "#" to "/auth/sign-up". This change ensures that clicking on the link takes the user to the sign-up page instead of scrolling to the top of the page.

Recent user commits:
- refactor: Update sign-up screen to display user's full name

Recent repository commits:
- refactor: Update Prisma schema to use UUIDs for ID fields
- refactor: Update sign-up screen and layout styles
- refactor: Update Prisma database provider to MongoDB
- fix:updated sign up screen
- refactor: Update database provider to MongoDB in Prisma schema
- feat: Add @prisma/client and mongodb dependencies
- addes transpile option in next js config
- added keys to the env files after testing