feat: Add lazy loading for images and improve course module loading
This commit adds lazy loading for images in the createMDX function of the next.config.mjs file. It also improves the loading of course modules in the LessonModule component by fetching the course data and serializing the contents. Additionally, it adds loading and navigation functionality for the sections of the course. These changes enhance the performance and user experience of the application.

Recent user commits:
- refactor: Update sign-up screen to use "/auth/sign-up" as the href
- update:created new course
- feat: Add @prisma/client and mongodb dependencies
- refactor: Update sign-up screen to display user's full name
- fix: Update status code to 400 for existing email error
- chore: Update package.json to include postinstall script for Prisma generation

Recent repository commits:
- refactor: Update sign-up screen to use "/auth/sign-up" as the href
- update:created new course
- feat: Add @prisma/client and mongodb dependencies
- refactor: Update sign-up screen to display user's full name
- fix: Update status code to 400 for existing email error
- chore: Update package.json to include postinstall script for Prisma generation
- refactor: Update Prisma schema to use UUIDs for ID fields